chore(config): rename eslint and stylelint config files
