+++ packages/config/src/eslint.cjs
