+++ packages/config/src/stylelint.cjs
